pyproject.toml
[project]
name = "gold-tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.2.0",
    "pytest>=9.0.1",
    "pytest-mock>=3.15.1",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "requests>=2.31.0",
    "sqlalchemy>=2.0.0",
]
[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
